backend/requirements.txt
fastapi
numpy
opencv-python
pillow
python-multipart
uvicorn==0.16.0